backend/requirements.txt
# Django Framework
Django==5.0.1
djangorestframework==3.14.0

# Authentication & Authorization
djangorestframework-simplejwt==5.3.1

# API Documentation
drf-yasg==1.21.7

# CORS Headers
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.10  # PostgreSQL adapter (comment if using SQLite)

# Environment Variables
python-decouple==3.8
setuptools>=70.0.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality
flake8==7.1.1
black==24.10.0

# Utilities
python-dateutil==2.8.2
Pillow==10.4.0

# Development
ipython==8.20.0
